Refactorizado el AuthProvider con otro valor para el usuario
