+++ src/auth/AuthPorvider.jsx
@@ -1,34 +1,4 @@
-import { createContext, useContext, useEffect, useState } from "react";
-
-import { createToken, getToken, deleteToken } from "../storage/sesiones";
-import { getUser, getUsers, getUserByUserName } from "../storage/users";
-
-export async function creandoUserToken(userName) {
-  const user = await getUserByUserName(userName);
-  const chainTokenPart1 = Math.random().toString(36).substring(2, 9);
-  const chainTokenPart2 = Math.random().toString(36).substring(2, 9);
-
-  const token = {
-    token: `${chainTokenPart1}${user.id}${chainTokenPart2}${user.username}`,
-    userid: user.id,
-  };
-
-  const newToken = await createToken(token);
-  if (!newToken) {
-    return false;
-  }
-
-  return token.token;
-}
-
-export async function borrarToken(idToken) {
-  const isDelete = await deleteToken(idToken);
-  if (isDelete) {
-    console.log("Se borro el token");
-  } else {
-    console.log("Error al borrar token");
-  }
-}
+import { createContext, useContext, useState } from "react";
 
 const AuthContext = createContext({
   isAuthenticated: false,
@@ -45,8 +15,26 @@ export function AuthProvider({ children }) {
     id: "error_id",
   });
 
+  const handleAuthenticat = (value) => {
+    let newAuth = !value == null && value;
+    if (newAuth) {
+      return { message: "Error al cambiar autenticacion" };
+    }
+    setIsAuthenticated(newAuth);
+  };
+
+  const handleUser = (value) => {
+    let newUser = !value == null && value;
+    if (newUser) {
+      return { message: "Error al cambiar valores de usuario" };
+    }
+    setUser(newUser);
+  };
+
   return (
-    <AuthContext.Provider value={{ isAuthenticated, user }}>
+    <AuthContext.Provider
+      value={{ isAuthenticated, user, handleAuthenticat, handleUser }}
+    >
       {children}
     </AuthContext.Provider>
   );
